create WorkerMapper & WorkerMapper.xml
