--- back/cctv/src/main/java/com/autoever/controller/delete_worker.java
@@ -1,4 +1,0 @@
-package com.autoever.controller;
-
-public class delete_worker {
-}
