--- back/cctv/src/main/java/com/autoever/controller/put_worker.java
@@ -1,4 +1,0 @@
-package com.autoever.controller;
-
-public class put_worker {
-}
