--- back/cctv/src/main/java/com/autoever/mapper/put_worker.java
@@ -1,4 +1,0 @@
-package com.autoever.mapper;
-
-public class put_worker {
-}
