+++ back/cctv/src/main/java/com/autoever/service/WorkerService.java
@@ -1,4 +1,6 @@
 package com.autoever.service;
 
-public class put_worker {
+
+
+public class WorkerService {
 }
