--- back/cctv/src/main/java/com/autoever/service/get_worker.java
@@ -1,4 +1,0 @@
-package com.autoever.service;
-
-public class get_worker {
-}
