--- back/cctv/src/main/java/com/autoever/service/update_worker.java
@@ -1,4 +1,0 @@
-package com.autoever.service;
-
-public class update_worker {
-}
